+++ CHANGES.md
@@ -21,21 +21,44 @@ To be released.
 
 ### Behavioral changes
 
+ -  `PolymorphicAction<T>` now allows subclasses of `T` to be declared in
+    the entry assembly of the application besides the same assembly as `T`.
+    [[#2136]]
+ -  (Libplanet.Net) Raised the default value for `Swarm<T>()` constructor's
+    `workers` parameter to 100.  [[#2128], [#2134]]
+ -  (Libplanet.Explorer) Raised the default value for `workers` console
+    argument to 100.  [[#2128], [#2134]]
+
 ### Bug fixes
+
+ -  Fixed a bug where `PrivateKey()` constructor had returned an invalid key
+    less than 32 bytes. [[#1696], [#2091]]
+ -  (Libplanet.Net) Invalid `Uri.UserInfo` with multiple colons is now
+    rejected by `IceServer(Uri url)` constructor and exception is thrown.
+    [[#2058], [#2116]]
 
 ### Dependencies
 
 ### CLI tools
 
- - (Libplanet.Extensions.Cocona) Upgraded *Cocona.Lite* from 1.6.\* to
+ -  (Libplanet.Extensions.Cocona) Upgraded *Cocona.Lite* from 1.6.\* to
     [2.0.\*][Cocona.Lite 2.0.0].  [[#2101]]
+ -  Implemented *planet key derive*, now you can get public key and
+    address from private key directly!  [[#2108]]
+ -  (Libplanet.Extensions.Cocona) Removed `DerivationCommand` class.
+    [[#2118]]
 
- - Implemented *planet key derive*, now you can get public key and
-   address from private key directly!  [[#2108]]
-
+[#1696]: https://github.com/planetarium/libplanet/issues/1696
+[#2058]: https://github.com/planetarium/libplanet/issues/2058
+[#2091]: https://github.com/planetarium/libplanet/pull/2091
 [#2101]: https://github.com/planetarium/libplanet/pull/2101
 [#2108]: https://github.com/planetarium/libplanet/pull/2108
+[#2116]: https://github.com/planetarium/libplanet/pull/2116
+[#2118]: https://github.com/planetarium/libplanet/pull/2118
+[#2128]: https://github.com/planetarium/libplanet/issues/2128
 [#2130]: https://github.com/planetarium/libplanet/pull/2130
+[#2134]: https://github.com/planetarium/libplanet/pull/2134
+[#2136]: https://github.com/planetarium/libplanet/pull/2136
 [Cocona.Lite 2.0.0]: https://www.nuget.org/packages/Cocona.Lite/2.0.0